requirements.txt
discord.py==2.2.2
openai==0.27.8
python-dotenv==0.20.0
langchain==0.0.198
google-api-python-client==2.89.0